Rewrite of documentation, and updated the Buildroot configuration files to use the latest 5.2 Linux kernels for 64-bit RISC-V targets and 4.19.66 for 32-bit RISC-V targets due to broken upstream support for RV32 in Linux 5.x

+++ src/hypervisor/loader.rs
@@ -84,7 +84,7 @@ pub fn load(target: Region, source: Region) -> Result<Entry, Cause>
         Some(e) => 
         {
             hvlog!("Supervisor kernel entry: 0x{:x}", e);
-            return Ok(unsafe { core::intrinsics::transmute(e) });
+            return Ok(unsafe { core::intrinsics::transmute(e as usize) });
         }
     }
 }
